move select instance query to middleware
