+++ apps/prairielearn/src/pages/studentAssessment/studentAssessment.ts
@@ -1,12 +1,10 @@
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
 
-import { AugmentedError, HttpStatusError } from '@prairielearn/error';
+import { HttpStatusError } from '@prairielearn/error';
 import { flash } from '@prairielearn/flash';
-import { loadSqlEquiv, queryOptionalRow } from '@prairielearn/postgres';
 
 import { makeAssessmentInstance } from '../../lib/assessment.js';
-import { IdSchema } from '../../lib/db-types.js';
 import {
   joinGroup,
   createGroup,
@@ -24,36 +22,10 @@ import { checkPasswordOrRedirect } from '../../middlewares/studentAssessmentAcce
 import { StudentAssessment } from './studentAssessment.html.js';
 
 const router = Router();
-const sql = loadSqlEquiv(import.meta.url);
 
 router.get(
   '/',
   asyncHandler(async function (req, res) {
-    if (res.locals.assessment.multiple_instance) {
-      if (res.locals.assessment.type === 'Homework') {
-        throw new AugmentedError('"Homework" assessments do not support multiple instances', {
-          data: { assessment: res.locals.assessment },
-        });
-      }
-      // The user has landed on this page to create a new assessment instance.
-      // Proceed even if there are existing instances.
-    } else {
-      // If the assessment is single-instance, check if the user already has an
-      // instance. If so, redirect to it.
-      const assessment_instance_id = await queryOptionalRow(
-        sql.select_single_assessment_instance,
-        {
-          assessment_id: res.locals.assessment.id,
-          user_id: res.locals.user.user_id,
-        },
-        IdSchema,
-      );
-      if (assessment_instance_id != null) {
-        res.redirect(`${res.locals.urlPrefix}/assessment_instance/${assessment_instance_id}`);
-        return;
-      }
-    }
-
     // Before allowing the user to create a new assessment instance, we need
     // to check if the current access rules require a password. If they do,
     // we'll ensure that the password has already been entered before allowing
